+++ CHANGELOG.md
@@ -65,6 +65,9 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 #### exonum-merkledb
 
+- Added `Debug` implementation for `Database`, `Snapshot`, `Iterator` dynamic
+  traits (#1363)
+
 - Changed storage layout (#1293)
 
   - Changed indexes metadata layout in the database.